use realpath in filename recognition.
So when you are editing a file that is a symbolic link to another file.
The filename is retrieved from the original file instead of the soft
link.

For example, you have a file named 'Thisisreal', and a symbolic link to
it named 'Thisisalink'. When you edit 'Thisisalink', in previous
versions of this plugin, the filename field in the header is
'Thisisalink'. But this commit would change this behavior.